Test injection with different DV plugin versions
Previously, our test coverage for injection of the DV plugin was limited to the latest DV plugin. The only exception was testing v3.16.2, but even then we failed to test setting the DV URL, capture fingerprints, etc.

With this change, we now test running the init-script with different DV versions to inject, ensuring our script continues to support injection of older plugin versions (which may be required if the DV server environment is out of date).