Test injection with different existing DV plugin versions
The init-script must be able to apply configuration to a Gradle build that already
has the DV plugin applied, and must not error regardless of the DV plugin version applied.

Previously, we only tested against Gradle builds using the latest DV plugin, and GE plugin v3.16.2.
With this change, we now test running the init-script on projects with many different DV versions applied.