`uploadInBackground` can be configured even Develocity plugin is already applied (#36)
Allow uploadInBackground to be configured when Develocity plugin already applied.

This is necessary for the Develocity Build Validation Scripts, but could also be useful for CI plugins.

Where this will change behavior for CI users is if the project is currently configuring uploadInBackground = true on CI. The CI plugins disable this if configured to do so.